+++ packages/demo/.eslintignore
@@ -1,2 +1,3 @@
 dist/
 node_modules/
+public/docs/
