+++ packages/demo/src/components/RightSideBar/RightSideBar.tsx
@@ -1,14 +1,15 @@
 import Drawer from "@mui/material/Drawer";
 import { useTheme } from "@mui/styles";
-import { useGlobalStyles } from "@src/styles";
 import { type PropsWithChildren } from "react";
 
-export const RightSideBar = ({ children }: PropsWithChildren<{}>): JSX.Element => {
+import { useGlobalStyles } from "@src/styles";
+
+export const RightSideBar = ({ children }: PropsWithChildren): JSX.Element => {
   const theme = useTheme();
   const classes = useGlobalStyles(theme);
 
   return (
-    <Drawer className={classes.rightSideBar} anchor="right" variant="permanent">
+    <Drawer anchor="right" className={classes.rightSideBar} variant="permanent">
       {children}
     </Drawer>
   );
