--- packages/demo/src/components/VerifiableCredentials/index.ts
@@ -1,3 +1,0 @@
-import { VerifiableCredentials } from "./VerifiableCredentials";
-
-export default VerifiableCredentials;
