+++ packages/demo/src/constants/paths.ts
@@ -1,11 +1,12 @@
 export enum Paths {
-    ROOT = "/",
-    HOME = "/home",
-    CONNECT = "/connect/:isChangeIdentityParam",
-    GET_IDENTITY_METADATA = "/get-connected-metadata",
-    IMPORT_IDENTITY =  "/import-identity",
-    REVEAL_IDENTITY_COMMITMENT = "/reveal-identity-commitment",
-    SEMAPHORE = "/semaphore",
-    RLN = "/rate-limiting-nullifier",
-    BANDADA = "/bandada"
+  ROOT = "/",
+  HOME = "/home",
+  CONNECT = "/connect/:isChangeIdentityParam",
+  GET_IDENTITY_METADATA = "/get-connected-metadata",
+  IMPORT_IDENTITY = "/import-identity",
+  REVEAL_IDENTITY_COMMITMENT = "/reveal-identity-commitment",
+  SEMAPHORE = "/semaphore",
+  RLN = "/rate-limiting-nullifier",
+  BANDADA = "/bandada",
+  VERIFIABLE_CREDENTIALS = "/verifiable-credentials",
 }
