+++ packages/demo/src/hooks/index.ts
@@ -1,2 +1,1 @@
-export { useCryptKeeper } from "./useCryptKeeper";
 export { useFileReader } from "./useFileReader";
