+++ packages/demo/src/pages/Bandada/index.ts
@@ -1,3 +1,3 @@
-import {lazy} from "react";
+import { lazy } from "react";
 
 export default lazy(() => import("./Bandada"));
