+++ packages/demo/src/pages/Connect/Connect.tsx
@@ -1,27 +1,26 @@
 import Box from "@mui/material/Box";
+import Container from "@mui/material/Container";
 import Typography from "@mui/material/Typography";
 import Markdown from "react-markdown";
-import Container from "@mui/material/Container";
-import rehypeSlug from "rehype-slug";
+import { useParams } from "react-router-dom";
 import rehypeAutolinkHeadings from "rehype-autolink-headings";
+import rehypeSlug from "rehype-slug";
 
 import ActionBox from "@src/components/ActionBox";
 import { useFileReader } from "@src/hooks";
 import { useGlobalStyles } from "@src/styles";
-import { useConnect } from "./useConnect";
-import { useParams } from "react-router-dom";
 
-type RouteParams = {
-  isChangeIdentityParam: string;
-};
+import { useConnect } from "./useConnect";
 
 const Connect = (): JSX.Element => {
   const classes = useGlobalStyles();
   const { connect } = useConnect();
-  const { isChangeIdentityParam } = useParams<RouteParams>();
+  const { isChangeIdentityParam } = useParams<{
+    isChangeIdentityParam?: string;
+  }>();
 
   const isChangeIdentity = isChangeIdentityParam === "true";
-  
+
   const { fileContent: code } = useFileReader("connect.ts");
   const { fileContent: doc } = useFileReader("connect.md");
 
@@ -31,7 +30,9 @@ const Connect = (): JSX.Element => {
         className={classes.popup}
         sx={{ p: 3, display: "flex", flexDirection: "column", alignItems: "center", justifyContent: "center" }}
       >
-        <Typography variant="h6">{isChangeIdentity ? "Connect to CryptKeeper" : "To continue, please connect to your CryptKeeper to continue."}</Typography>
+        <Typography variant="h6">
+          {isChangeIdentity ? "Connect to CryptKeeper" : "To continue, please connect to your CryptKeeper to continue."}
+        </Typography>
 
         <Markdown rehypePlugins={[rehypeSlug, [rehypeAutolinkHeadings, { behavior: "wrap" }]]}>{doc}</Markdown>
