+++ packages/demo/src/pages/RateLimitingNullifier/index.ts
@@ -1,3 +1,3 @@
-import {lazy} from "react";
+import { lazy } from "react";
 
 export default lazy(() => import("./RateLimitingNullifier"));
