+++ packages/demo/src/pages/VerifiableCredentials/docs/addVC.ts
