+++ packages/demo/src/pages/VerifiableCredentials/docs/generateVP.ts
