+++ packages/demo/src/pages/VerifiableCredentials/index.ts
@@ -1,0 +1,3 @@
+import { lazy } from "react";
+
+export default lazy(() => import("./VerifiableCredentials"));
