+++ packages/demo/webpack.dev.js
@@ -1,4 +1,5 @@
 const { merge } = require("webpack-merge");
+
 const path = require("path");
 
 const common = require("./webpack.config");
